+++ app/scheduler/tasks/export_definitions/config_scraper/__init__.py
@@ -1,0 +1,2 @@
+from .config_scraper_xls import XlsExportConfig
+from .config_scraper_shp import ShpExportConfig
