+++ app/scheduler/tasks/export_definitions/export/export_xls.py
@@ -1,5 +1,3 @@
-import math
-import logging
 import pathlib
 import openpyxl
 
@@ -7,65 +5,27 @@ from datetime import datetime
 from openpyxl.utils import cell
 
 from django.conf import settings
-from django.db import connection, connections, ProgrammingError
+from django.db import connections, ProgrammingError
 
-from .domains_parser import Domains
-from .config_scraper import ExportConfig
-
-from app.scheduler.models import Task
 from app.scheduler.utils import dictfetchall, translate_schema_to_db_alias
+from app.scheduler.tasks.export_definitions.domains_parser import Domains
+from app.scheduler.tasks.export_definitions.config_scraper import XlsExportConfig
+
+from .export_base import ExportBase
 
 
-class ExportXls:
+class ExportXls(ExportBase):
 
     SEED_FILE_ID_ROW = 3
-
-    def __init__(
-        self, export_dir: pathlib.Path, orm_task: Task, max_progress: int = 100
-    ):
-        """
-        Initialization function of data export to *.xlsx file
-
-        Parameters:
-            export_dir: directory where export and log files should be stored
-            orm_task: instance of the database Task reporting execution of this export task (default 100)
-            max_progress: max value of Task's progress, which should be set after a successful export
-        """
-        self.export_dir = export_dir
-        self.orm_task = orm_task
-        self.max_progress = max_progress
-
-        # make sure target location exists
-        self.export_dir.parent.mkdir(parents=True, exist_ok=True)
-
-    def configure_file_logger(self, logfile_path: pathlib.Path):
-        """
-        Method configuring logger for logging user dedicated errors of the export into a specified location.
-
-        Parameters:
-            logfile_path: path of the log file
-        """
-        logger = logging.getLogger(__name__)
-        hdlr = logging.FileHandler(logfile_path.absolute())
-        formatter = logging.Formatter("%(asctime)s %(levelname)s %(message)s")
-        hdlr.setFormatter(formatter)
-        logger.addHandler(hdlr)
-        logger.setLevel(logging.INFO)
-
-        return logger
 
     def run(self):
         """
         Method executing export of the data into *.xlsx file,
         """
-        starting_progress = self.orm_task.progress
+        # update starting progress
+        self.starting_progress = self.orm_task.progress
         today = datetime.today()
-
-        # create log file
-        logfile_path = pathlib.Path(
-            self.export_dir.absolute(), f"logfile_{today.strftime('%Y%m%d')}.log"
-        )
-        logger = self.configure_file_logger(logfile_path)
+        self.configure_file_logger()
 
         # get target xls file location
         target_xls_file = pathlib.Path(
@@ -74,7 +34,7 @@ class ExportXls:
         )
 
         # parse export configuration
-        config = ExportConfig()
+        config = XlsExportConfig()
 
         # calculate total number of steps
         total_sheet_number = len(config)
@@ -86,32 +46,25 @@ class ExportXls:
         # load seed *.xlsx file
         excel_wb = openpyxl.load_workbook(settings.EXPORT_XLS_SEED_FILE.substitute())
 
-        for sheet in ExportConfig():
+        for sheet in config:
 
             # execute pre_process for the sheet
             pre_process = sheet.get("pre_process", None)
-            if pre_process is not None:
-                with connection.cursor() as cursor:
-                    try:
-                        cursor.callproc(f"{self.orm_task.schema}.{pre_process}")
-                    except Exception as e:
-                        logger.error(
-                            f"Procedure '{pre_process}' called by sheet '{sheet['sheet']}' FAILED with:\n"
-                            f"{type(e).__name__}: {e}.\n"
-                            f"Skipping '{sheet['sheet']}' sheet generation."
-                        )
-                        continue
-                    else:
-                        result = cursor.fetchone()
-                        logger.debug(
-                            f"Procedure '{pre_process}' called by sheet '{sheet['sheet']}' executed: {result}."
-                        )
+            try:
+                self.execute_pre_process(pre_process)
+            except Exception as e:
+                self.logger.error(
+                    f"Procedure '{pre_process}' called by sheet '{sheet['sheet']}' FAILED with:\n"
+                    f"{type(e).__name__}: {e}.\n"
+                    f"Skipping '{sheet['sheet']}' sheet generation."
+                )
+                continue
 
             # set current sheet as active in the xls workbook
             try:
                 sheet_index = excel_wb.sheetnames.index(sheet["sheet"])
             except ValueError:
-                logger.error(
+                self.logger.error(
                     f"Seed file does not contain '{sheet['sheet']}' sheet. Skipping..."
                 )
                 continue
@@ -141,7 +94,7 @@ class ExportXls:
                 sql_sources = sheet["sql_sources"]
 
                 if not sql_sources:
-                    logger.warning(
+                    self.logger.warning(
                         f"Sources for '{sheet['sheet']}' is empty. Skipping..."
                     )
                     continue
@@ -151,7 +104,7 @@ class ExportXls:
                     try:
                         cursor.execute(source)
                     except ProgrammingError as e:
-                        logger.error(
+                        self.logger.error(
                             f"Fetching source for sheet '{sheet['sheet']}' failed with:\n"
                             f"{type(e).__name__}: {e}.\n"
                             f"Source: '{source}'."
@@ -170,7 +123,7 @@ class ExportXls:
                             row=raw_data_row, domains=all_domains
                         )
                     except Exception as e:
-                        logger.error(
+                        self.logger.error(
                             f"Error occurred during transformation of column with "
                             f"ID '{column['id']}' in row '{first_empty_row}' in sheet '{sheet['sheet']}':\n"
                             f"{type(e).__name__}: {e}.\n"
@@ -192,9 +145,9 @@ class ExportXls:
                                     .replace("{ROW}", first_empty_row)
                                     .replace("{FIELD}", column_letter)
                                 )
-                                logger.error(message)
+                                self.logger.error(message)
                             else:
-                                logger.error(
+                                self.logger.error(
                                     f"Validation failed for cell '{column_letter}{first_empty_row}' "
                                     f"in the '{sheet['sheet']}' sheet."
                                 )
@@ -203,7 +156,7 @@ class ExportXls:
                 for column_id, value in sheet_row.items():
                     column_letter = coord_id_mapping.get(str(column_id))
                     if not column_letter:
-                        logger.error(
+                        self.logger.error(
                             f"No column with ID '{column_id}' found in '{sheet['sheet']}' sheet."
                         )
                         continue
@@ -211,7 +164,7 @@ class ExportXls:
                     try:
                         excel_ws[f"{column_letter}{first_empty_row}"] = value
                     except Exception as e:
-                        logger.error(
+                        self.logger.error(
                             f"Error occurred during inserting value to the column with "
                             f"ID '{column_id}' in row '{first_empty_row}' in sheet '{sheet['sheet']}':\n"
                             f"{type(e).__name__}: {e}.\n"
@@ -221,13 +174,9 @@ class ExportXls:
 
             # update task status
             step += 1
-            self.orm_task.progress = math.floor(
-                step * (self.max_progress - starting_progress) / total_sheet_number
-            )
-            self.orm_task.save()
+            self.update_progress(step, total_sheet_number)
 
         # save updated *.xlsx seed file in the target location
         excel_wb.save(target_xls_file)
 
-        self.orm_task.progress = self.max_progress
-        self.orm_task.save()
+        self.set_max_progress()
