+++ export/config/current/shapefile_configs/config_fiumi.json
@@ -1,0 +1,54 @@
+{
+    "file_name": "fiumi.shp",
+    "skip": false,
+    "folder": "Impianti",
+    "sources": [
+        {
+            "fields": [
+                {"name":"acq_captazione.geom", "alias": "geom"},
+                {"name":"acq_captazione.geom", "alias": "x_proj", "function": "GB_X"},
+                {"name":"acq_captazione.geom", "alias": "y_proj", "function": "GB_Y"},
+                {"name":"acq_captazione.codice_ato","alias": "codice_ato"}
+            ],
+            "table": {
+                "name": "acq_captazione"
+            },
+            "join": [
+                {
+                    "type": "inner",
+                    "table": {
+                        "name": "acq_auth_capt"
+                    },
+                    "on": [
+                        "acq_captazione.idgis",
+                        "acq_auth_capt.id_captazione"
+                    ],
+                    "cond": "="
+                },
+                {
+                    "type": "inner",
+                    "table": {
+                        "name": "acq_capt_conces"
+                    },
+                    "on": [
+                        "acq_captazione.idgis",
+                        "acq_capt_conces.id_captazione"
+                    ],
+                    "cond": "="
+                },
+                {
+                    "type": "inner",
+                    "table": {
+                        "name": "localita"
+                    },
+                    "on": [
+                        "localita.geom",
+                        "acq_captazione.geom"
+                    ],
+                    "cond": "ST_INTERSECTS"
+                }
+            ],
+            "filter": "WHERE acq_captazione.d_gestore = 'PUBLIACQUA' AND acq_captazione.d_ambito IN ('AT3', NULL) AND acq_captazione.d_stato NOT IN ('IPR', 'IAC') and SUB_FUNZIONE=0"
+        }
+    ]
+}
