+++ export/config/current/xls.json
@@ -1,29 +1,5 @@
 {
    "xls_sheet_configs": [
-      "sheet_configs/config_accumuli.json",
-      "sheet_configs/config_addut_tronchi.json",
-      "sheet_configs/config_adduttrici.json",
-      "sheet_configs/config_collett_tronchi.json",
-      "sheet_configs/config_collettori.json",
-      "sheet_configs/config_depurato_pompe.json",
-      "sheet_configs/config_depuratori.json",
-      "sheet_configs/config_distrib_tronchi.json",
-      "sheet_configs/config_distribuzioni.json",
-      "sheet_configs/config_potabilizzatori.json",
-      "sheet_configs/config_pozzi.json",
-      "sheet_configs/config_pozzi_pompe.json",
-      "sheet_configs/config_scaricatori.json",
-      "sheet_configs/config_fiumi.json",
-      "sheet_configs/config_fognat_com_serv.json",
-      "sheet_configs/config_fognat_loc_serv.json",
-      "sheet_configs/config_fognature.json",
-      "sheet_configs/config_laghi.json",
-      "sheet_configs/config_pompaggi.json",
-      "sheet_configs/config_pompaggi_pompe.json",
-      "sheet_configs/config_potab_pompe.json",
-      "sheet_configs/config_fognat_tronchi.json",
-      "sheet_configs/config_sollev_pompe.json",
-      "sheet_configs/config_sollevamenti.json",
-      "sheet_configs/config_sorgenti.json"
+      "sheet_configs/config_fiumi.json"
    ]
 }
