+++ requirements.txt
@@ -23,3 +23,5 @@ pypika==0.43.0
 schema==0.7.2
 openpyxl==3.0.5
 django-postgres-copy==2.4.2
+shapely==1.7.1
+fiona==1.8.18
